requirements_bitsproject.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.0.5
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
dataclasses-json==0.6.4
datasets==2.18.0
debugpy==1.6.0
decorator==5.1.1
dill==0.3.8
distro==1.9.0
dotenv-python==0.0.1
entrypoints==0.4
exceptiongroup==1.2.0
executing==0.8.3
fastapi==0.110.0
fastapi-pagination==0.12.7
filelock==3.13.3
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.22.0
idna==3.6
ipykernel==6.15.0
ipython==8.4.0
jedi==0.18.1
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-client==7.3.4
jupyter-core==4.10.0
langchain==0.1.13
langchain-community==0.0.29
langchain-core==0.1.34
langchain-openai==0.1.1
langchain-text-splitters==0.0.1
langsmith==0.1.33
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib-inline==0.1.3
mdurl==0.1.2
mmh3==4.1.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.5.5
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
openai==1.14.3
orjson==3.9.15
packaging==23.2
pandas==2.2.1
parso==0.8.3
pdfminer.six==20231228
pickleshare==0.7.5
pillow==10.2.0
pinecone-client==3.2.1
pinecone-text==0.9.0
prompt-toolkit==3.0.29
protobuf==4.25.3
psutil==5.9.1
pure-eval==0.2.2
pyarrow==15.0.2
pyarrow-hotfix==0.6
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
pywin32==304
PyYAML==6.0.1
pyzmq==23.1.0
rank-bm25==0.2.2
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==2.6.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.29
stack-data==0.3.0
starlette==0.36.3
streamlit==1.32.2
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.4.0
tiktoken==0.6.0
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.2.2+cu118
tornado==6.1
tqdm==4.66.2
traitlets==5.3.0
transformers==4.39.2
types-requests==2.31.0.20240311
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
watchdog==4.0.0
wcwidth==0.2.5
wget==3.2
xxhash==3.4.1
yarl==1.9.4
